refactor: 테스트용 가짜 객체들 패키지 분리
