+++ src/test/java/kitchenpos/domain/InMemoryMenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.domain;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -6,8 +6,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuGroupRepository;
 
 public class InMemoryMenuGroupRepository implements MenuGroupRepository {
