+++ src/test/java/kitchenpos/domain/InMemoryMenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.domain;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -8,8 +8,6 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuRepository;
 
 public class InMemoryMenuRepository implements MenuRepository {
