+++ src/test/java/kitchenpos/domain/InMemoryOrderRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.domain;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -7,10 +7,6 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderRepository;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
 
 public class InMemoryOrderRepository implements OrderRepository {
