+++ src/test/java/kitchenpos/domain/InMemoryOrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.domain;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -6,8 +6,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.OrderTableRepository;
 
 public class InMemoryOrderTableRepository implements OrderTableRepository {
